pyproject.toml
[tool.poetry]
name = "vinylizer"
version = "0.1.0"
description = ""
authors = ["alice"]
license = "GPL"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
python3-discogs-client = "^2.7"
tomli = "^2.0.1"
pyside6 = "6.6.1"
pillow = "^10.1.0"
python-slugify = "^8.0.1"
openpyxl = "^3.1.2"


[tool.poetry.group.dev.dependencies]
pynvim = "^0.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
